pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bid-review"
version = "0.1.0"
description = "Tender/Bid compliance review pipeline driven by local Claude Code CLI."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
